+++ CHANGES.md
@@ -11,7 +11,8 @@ Unreleased
   operation with unwritten bytes (#103, #116)
 - Fix calls in C stubs that need to call `ERR_clear_error` before the underlying
   OpenSSL call (#118)
-
+- Add a module `Ssl.Error` to retrieve OpenSSL errors in a structured way (#119)
+  
 0.5.13 (2022-10-20)
 =====